Something that actually almost works!
There are a lot more features and things that I need to add, but I have
the skeleton of a UI that allows a user to pop open a bunch of chat
windows. Next up, I have to perfect this, and then work on getting a
concept of users integrated.